Disable max-statements
This rule falls into a trap. Because the person who first triggers the warning
is not necessarily individually responsible for causing it, _and_ fixing this
kind of issue can be non-trivial, these warnings tend to go unfixed.

To reduce warning noise, and thus increase the value of other warnings,
I propose we remove this warning.